Fix crash when stream buffer outlives the sent_req
Because of the timeout on bt request streams, it's possible for the
`sent_request` object that owns the request data to get destroyed before
the data is acked; subsequent attempts to send will then be referencing
invalid memory because the `send()` was only using a view with no
ownership.

This transfers the data into the `send()` call so that the stream
buffers keep it alive as long as needed, and it doesn't seem like
keeping `data` around in the sent_req is needed for anything.