Fix when_writeable firing after a connection has gone away
If a connection send would block then we set up a when_writeable timer,
but this could fire *after* the Connection has been destroyed, at which
point we'd be running with `this` not being valid.

This fixes it to make sure the endpoint still knows about the connection
before we try to re-send.